# src/app/api/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentFoundry API",
    "description": "ERC-8004 agent registration and discovery API",
    "version": "0.1.0"
  },
  "servers": [{ "url": "/", "description": "Current host" }],
  "paths": {
...
    "/api/v1/webhooks": {
      "post": {
        "summary": "Register webhook",
        "description": "Register URL for events: registration, uri_update, feedback, feedback_revoked, ownership_transfer, validation_response. Requires API key when configured.",
        "tags": ["Webhooks"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "events"],
                "properties": {
                  "url": { "type": "string", "format": "uri" },
                  "events": { "type": "array", "items": { "type": "string", "enum": ["registration", "uri_update", "feedback", "feedback_revoked", "ownership_transfer", "validation_response"] } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Webhook registered" } }
      }
    },
    "/api/v1/events": {
      "post": {
        "summary": "Emit event (internal)",
        "description": "Triggers webhook delivery. Called by frontend after registration, uri_update, feedback, transfer, revoke.",
        "tags": ["Webhooks"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event": { "type": "string", "enum": ["registration", "uri_update", "feedback", "feedback_revoked", "ownership_transfer", "validation_response"] },
                  "data": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}